third_party/blink/web_tests/platform/generic/fast/frames/iframe-onload-and-domnodeinserted-expected.txt

This test ensures that any tree mutation in the load event handler cannot harm the tree consistency.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS loadEventFired is true
PASS unless crash.
PASS successfullyParsed is true

TEST COMPLETE
